refactor(LottoController, LottoInput, InteractionRepeatable): 재입력의 책임을 컨트롤러로 이동
- 재입력 판단 책임이 `뷰`의 범위를 넘어서므로, 전체를 핸들링 할 수 있는 컨트롤러에서 재입력의 책임을 수행하도록 책임 이동
- `재입력가능` 인터페이스의 패키지를 뷰에서 컨트롤러로 이동
