+++ src/main/java/lotto/view/LottoInput.java
@@ -1,32 +1,26 @@
 package lotto.view;
 
 import lotto.common.view.ConsoleInput;
-import lotto.common.view.InteractionRepeatable;
 import lotto.model.WinningLotto;
 import lotto.value.BonusNumber;
 import lotto.value.LottoNumbers;
 import lotto.value.Won;
 
-public class LottoInput implements InteractionRepeatable {
+public class LottoInput {
 
     public Won askForPurchasePrice() {
-        return supplyWithTry(() -> {
-            System.out.println("구입금액을 입력해 주세요.");
-            return Won.of(ConsoleInput.readInteger());
-        });
+        System.out.println("구입금액을 입력해 주세요.");
+        return Won.of(ConsoleInput.readInteger());
     }
 
     public WinningLotto askForWinningLotto() {
-        LottoNumbers winningNumbers = supplyWithTry(() -> {
-            System.out.println("당첨 번호를 입력해 주세요.");
-            return LottoNumbers.of(ConsoleInput.readIntegers());
-        });
+        System.out.println("당첨 번호를 입력해 주세요.");
+        LottoNumbers winningNumbers = LottoNumbers.of(ConsoleInput.readIntegers());
 
-        return supplyWithTry(() -> {
-            System.out.println("보너스 번호를 입력해 주세요.");
-            BonusNumber bonusNumber = new BonusNumber(ConsoleInput.readInteger());
-            return new WinningLotto(winningNumbers, bonusNumber);
-        });
+        System.out.println("보너스 번호를 입력해 주세요.");
+        BonusNumber bonusNumber = new BonusNumber(ConsoleInput.readInteger());
+
+        return new WinningLotto(winningNumbers, bonusNumber);
     }
 
 }
